mov status class
