+++ backend/apps/mentorship/models/common/__init__.py
@@ -1,4 +1,3 @@
 from .experience_level import ExperienceLevel
 from .matching_attributes import MatchingAttributes
 from .start_end_range import StartEndRange
-from .status import Status
